requirements.txt
langchain
google-generativeai
PyMuPDF
requests
python-dotenv
streamlit
